+++ CHANGELOG.md
@@ -89,6 +89,7 @@ Version 1.5.0
 - fix possible buffer overflow in console input buffer (RF bug)
 - fix possible server side crash when kicking a player using rcon (RF bug)
 - fix alpha draw order issues in Hendrix, Generator_Small01, Generator_Small02, LavaTester01, Weapon_RiotShield meshes (RF bug)
+- fix corona lights being visible through clutter objects (RF bug)
 
 Version 1.4.1 (released 2019-12-07)
 -----------------------------------